common guard code is now inherited from
