+++ transcendence-app/src/authorization/guards/base-policies.guard.ts
@@ -2,8 +2,8 @@ import { AnyMongoAbility } from '@casl/ability';
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { CaslAbilityFactory } from '../casl-ability.factory';
-import { CHECK_POLICIES_KEY } from '../decorators/policies.decorator';
 import { AuthorizationService } from '../authorization.service';
+import { CHECK_POLICIES_KEY } from '../decorators/policies.decorator';
 
 export interface IPolicyHandler {
   handle(ability: AnyMongoAbility): boolean;
@@ -14,35 +14,30 @@ export type PolicyHandlerCallback = (ability: AnyMongoAbility) => boolean;
 export type PolicyHandler = IPolicyHandler | PolicyHandlerCallback;
 
 @Injectable()
-export class CrMemberPoliciesGuard implements CanActivate {
-  constructor(
-    private reflector: Reflector,
-    private caslAbilityFactory: CaslAbilityFactory,
-    private readonly authorizationService: AuthorizationService,
+export abstract class PoliciesGuard implements CanActivate {
+  protected constructor(
+    protected reflector: Reflector,
+    protected caslAbilityFactory: CaslAbilityFactory,
+    protected readonly authorizationService: AuthorizationService,
   ) {}
 
-  async canActivate(ctx: ExecutionContext): Promise<boolean> {
+  abstract getAbilityFromContext(
+    ctx: ExecutionContext,
+  ): Promise<AnyMongoAbility>;
+
+  public async canActivate(ctx: ExecutionContext): Promise<boolean> {
     const policyHandlers =
       this.reflector.get<PolicyHandler[]>(
         CHECK_POLICIES_KEY,
         ctx.getHandler(),
       ) || [];
 
-    const chatId = ctx.switchToHttp().getRequest().params['chatroomId'];
-    const authId = ctx.switchToHttp().getRequest().user.id;
-    const authCrm =
-      await this.authorizationService.getUserAuthContextForChatroom(
-        authId,
-        chatId,
-      );
-
-    const ability = await this.caslAbilityFactory.defineAbilitiesFor(authCrm);
+    const ability = await this.getAbilityFromContext(ctx);
 
     return policyHandlers.every((handler) =>
       this.execPolicyHandler(handler, ability),
     );
   }
-
   private execPolicyHandler(handler: PolicyHandler, ability: AnyMongoAbility) {
     if (typeof handler === 'function') {
       return handler(ability);
